# docs/API_Contract_260714.yaml
openapi: 3.1.0
info:
  title: HUG 안심전세 체인 API Contract
  version: "1.0.0"
  summary: >-
    Frontend / Backend / ML / Blockchain 공통 계약(Single Source of Truth) 문서.
    Backend 구현 문서가 아니며 FastAPI 코드도 아니다.
  description: |
    ## 문서 성격
    본 문서는 HUG × 아이엔 안심전세 체인(HUG 안심전세 체인) 프로젝트의 **Frontend / Backend / ML / Blockchain 공통 계약**이다.
    Backend를 새로 설계하는 문서가 아니라, 이미 확정된 4개 상위 문서(개발설계보고서, 데이터수집 및 API가이드, ML개발가이드,
    Frontend UI/UX 명세서)를 근거로 프론트엔드가 axios 호출을 만들고, 백엔드가 FastAPI를 구현하고, ML Notebook이
    inference endpoint를 구현할 수 있도록 하는 **OpenAPI 3.1 단일 기준선**이다.

    ## 서비스 한 줄 요약
    HUG 안심전세 체인은 위험을 점수화하는 앱이 아니라, 계약 전 위험을 발견하고 위험조건이 실제로 해소되었는지
    증빙과 외부 API로 검증하며, 계약서·검증 이력을 블록체인에 공증하고, 계약 만료 전 D-90 반환계획 확인과
    사고 후 HUG 채권관리 인계까지 하나의 생애주기로 연결하는 주거안전 인프라다.

    ## 선행 문서와 우선순위
    | 영역 | 기준 문서 | 충돌 시 우선순위 |
    |---|---|---|
    | 공통 식별자·상태 | 개발설계보고서_260714_수정보완.docx 10장, 부록 B·C | 최우선 |
    | 외부 API/Mock | 데이터수집_및_API가이드_260714.md | 외부 API 상세 |
    | ML 입출력 계약 | ML개발가이드_260714.md 20장, 부록 C | 모델 Request/Response |
    | 화면·Route·API Mapping | Frontend_UIUX_명세서_260714.md 2·5·8장 | 화면-API 연결 |

    본 문서(API_Contract_260714.yaml)는 위 4개 문서의 API 관련 내용을 통합한 **실행 가능한 계약**이며,
    향후 작성될 Backend_API_명세서_260714.md, Blockchain_설계서_260714.md는 본 계약과 충돌할 수 없다.

    ## 데이터 신뢰성에 대한 명시 (ML개발가이드 2·30장 계승)
    - 정상계약(사고 없이 종료된 계약) 데이터가 없어 "실제 사고확률"을 추정하는 모델은 존재하지 않는다.
    - HUG 사고·경매·배당 데이터 7종은 전량 합성데이터이며, 회수·처리기간 관련 응답은 "합성데이터 기준 패턴"이다.
...

    ## 온체인 금지 원칙 (개발설계보고서 8장)
...

    ## 문서 요약 통계
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      operationId: authLogin
      tags: [Auth]
      summary: 로그인
      description: |
        이메일/비밀번호 기반 로그인 후 JWT Access Token과 역할(role) 클레임을 발급한다.
        발급된 role은 tenant, landlord, advisor, hug_admin, system_admin, verifier 중 하나이며
        Frontend RoleGuard가 이 값을 기준으로 라우트 접근을 제어한다(Frontend_UIUX_명세서 3장).
      security: []
      x-permission: 없음 (비로그인 사용자 전용)
      x-mock-supported: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: tenant01@example.com
              password: "P@ssw0rd!"
      responses:
        "200":
          description: 로그인 성공
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                status: Success
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type: Bearer
                  expires_in: 3600
                  user:
                    user_id: "8f14e45f-ceea-4b4f-9c1a-000000000001"
                    role: tenant
                    display_name: "홍길동"
                request_id: "req_20260714_0001"
        "422":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "408":
          $ref: '#/components/responses/Timeout'
  /auth/me:
    get:
      operationId: authMe
      tags: [Auth]
      summary: 내 세션 정보 조회
      description: |
        Bearer 토큰으로 현재 로그인한 사용자 정보(user_id, role, display_name, 최근 로그인 시각)를 조회한다.
        Frontend `useSessionStore` 초기화 시 호출한다(Frontend_UIUX_명세서 13.2절).
      x-permission: 로그인한 전체 역할 (tenant, landlord, advisor, hug_admin, system_admin, verifier)
      x-mock-supported: true
      responses:
        "200":
          description: 세션 정보 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              example:
                status: Success
                data:
                  user_id: "8f14e45f-ceea-4b4f-9c1a-000000000001"
                  role: tenant
                  display_name: "홍길동"
                  email: "tenant01@example.com"
                  created_at: "2026-06-01T09:00:00+09:00"
                  last_login_at: "2026-07-14T09:12:00+09:00"
                request_id: "req_20260714_0002"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /auth/logout:
    post:
      operationId: authLogout
      tags: [Auth]
      summary: 로그아웃
      description: 현재 세션의 Access Token을 서버 측에서 무효화한다. Refresh Token 정책은 Backend_API_명세서_260714.md에서 확정한다.
      x-permission: 로그인한 전체 역할
      x-mock-supported: true
      responses:
        "200":
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: Success
                data: { logged_out: true }
                request_id: "req_20260714_0003"
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /blockchain/anchor:
    post:
      operationId: anchorBlockchain
      tags: [Blockchain]
      summary: 결과 해시 온체인 기록
      description: |
        검증 완료, 계약버전 확정, D-90 이벤트, 사고 인계 등 감사 대상 사건의 canonical JSON 해시를 Polygon Amoy
        또는 Mock chain에 기록한다(개발설계보고서 8·9.7절). 온체인에는 case_id_hash, result_hash,
        model_version_hash, created_at, issuer만 저장하며 원문·개인정보는 절대 포함하지 않는다.
      x-permission: Backend 내부 서비스 호출 전용 (Contract/Verification/Scheduler/Incident Service). 사용자 직접 호출 대상 아님.
      x-mock-supported: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnchorRequest'
            example:
              event_type: VerificationCompleted
              reference_id: "ee000000-0000-0000-0000-000000000001"
              result_hash: "sha256:2c624232cdd221771294dfbb310aca000a0df6ac8b66b696d90ef06fdefb64a"
              model_version: null
      responses:
        "202":
          description: 온체인 기록 요청 접수(BlockchainStatus=Pending)
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestIdHeader'
            X-Blockchain-Version:
              $ref: '#/components/headers/BlockchainVersionHeader'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AnchorResponse'
              example:
                status: Success
                data:
                  blockchain_tx_id: "ff000000-0000-0000-0000-000000000001"
                  chain_id: 80002
                  contract_address: "0x0000000000000000000000000000000000dEaD"
                  tx_hash: null
                  blockchain_status: Pending
                  is_mock: false
                request_id: "req_20260714_0027"
        "422":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "408":
          $ref: '#/components/responses/Timeout'
  /blockchain/{tx_id}:
    get:
      operationId: getBlockchainTransaction
      tags: [Blockchain]
      summary: 블록체인 트랜잭션 상세 조회
      description: |
        blockchain_tx_id 기준 온체인 트랜잭션 상태(Pending/Confirmed/Failed)와 tx_hash, contract_address를 조회한다.
        `/blockchain/[txId]` 공통 화면과 TEN-05, ADM-02에서 사용한다. 법적 효력이 있는 것처럼 표현하지 않는다.
      x-permission: 로그인한 전체 역할(본인 관련 트랜잭션 또는 system_admin 전체)
      x-mock-supported: true
      parameters:
        - name: tx_id
          in: path
          required: true
          description: blockchain_tx_id (UUID) 또는 온체인 tx_hash
          schema:
            type: string
      responses:
        "200":
          description: 블록체인 트랜잭션 조회 성공
          headers:
            X-Blockchain-Version:
              $ref: '#/components/headers/BlockchainVersionHeader'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BlockchainTransaction'
              example:
                status: Success
                data:
                  blockchain_tx_id: "ff000000-0000-0000-0000-000000000001"
                  event_type: VerificationCompleted
                  reference_id: "ee000000-0000-0000-0000-000000000001"
                  chain_id: 80002
                  contract_address: "0x0000000000000000000000000000000000dEaD"
                  tx_hash: "0x9a3f...c21e"
                  result_hash: "sha256:2c624232cdd221771294dfbb310aca000a0df6ac8b66b696d90ef06fdefb64a"
                  blockchain_status: Confirmed
                  is_mock: false
                  created_at: "2026-07-14T15:20:00+09:00"
                  confirmed_at: "2026-07-14T15:20:45+09:00"
                request_id: "req_20260714_0028"
        "404":
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer 토큰. `role` 클레임(tenant/landlord/advisor/hug_admin/system_admin/verifier)을 포함한다.
        `verifier`는 advisor 내 서브 권한으로, `/verifications/{id}/decision` 액션 권한 판정에만 사용한다
        (Frontend_UIUX_명세서 3장, 개발설계보고서 13.5절). 발급/폐기 상세 구현은 Backend_API_명세서_260714.md에서 확정한다.
...
  headers:
    RequestIdHeader:
      description: 요청-응답을 연결하는 추적 ID. 클라이언트가 보낸 `Request-ID`를 그대로 반환하거나 서버가 생성한다.
      schema:
        type: string
...
    BlockchainVersionHeader:
      description: 응답 생성에 사용된 스마트컨트랙트/ABI 버전(블록체인 API에서만 반환).
      schema:
        type: string
  responses:
...
    ValidationError:
      description: 요청 값 검증 실패(422). Zod/Pydantic 검증 오류를 필드 단위로 반환한다.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      error_code:
                        type: string
                        example: ERROR-002
                      message:
                        type: string
                        example: "요청 값이 유효하지 않습니다."
                      field_errors:
                        type: array
                        items:
                          $ref: '#/components/schemas/ValidationErrorItem'
          example:
            status: Failed
            error:
              error_code: ERROR-002
              message: "요청 값이 유효하지 않습니다."
              field_errors:
                - field: deposit
                  reason: "0보다 큰 정수여야 합니다."
            request_id: "req_20260714_0100"
    Unauthorized:
      description: 인증 실패(401). 토큰 없음/만료/서명 불일치.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-004, message: "인증 토큰이 유효하지 않습니다." }
            request_id: "req_20260714_0101"
...
    NotFound:
      description: 리소스 없음(404). 존재하지 않거나 본인 소유가 아닌 리소스에 대한 조회.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-006, message: "요청한 리소스를 찾을 수 없습니다." }
            request_id: "req_20260714_0103"
...
    Timeout:
      description: 외부 API/모델 응답 지연(408). 데이터수집_및_API가이드 5장 Fallback Flow 대상.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-008, message: "외부 API 응답이 지연되어 재시도가 필요합니다." }
            request_id: "req_20260714_0105"
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      description: 전역 고유 식별자. 개발설계보고서 10.1절 공통 식별자 16종은 모두 이 포맷을 사용한다.
      example: "8f14e45f-ceea-4b4f-9c1a-000000000001"
    Timestamp:
      type: string
      format: date-time
      description: KST 기준 ISO 8601 일시(개발설계보고서 10.5절). 오프셋 `+09:00`을 포함해 표기한다.
      example: "2026-07-14T13:30:00+09:00"
...
    SuccessResponse:
      type: object
      description: 모든 성공 응답의 공통 봉투(envelope). data 필드는 각 API별 실제 스키마로 대체된다.
      required: [status, data, request_id]
      properties:
        status:
          type: string
          enum: [Success, Partial]
          example: Success
        data:
          description: 실제 응답 페이로드. 각 API operation에서 구체 스키마로 재정의한다.
          type: object
        request_id:
          type: string
          description: Request-ID 헤더와 동일한 값을 본문에도 포함해 로그 상관관계를 보장한다.
          example: "req_20260714_0001"
    ErrorResponse:
      type: object
      description: 모든 오류 응답의 공통 봉투. error_code는 "ERROR-NNN" 형식(하단 오류코드 표 참조).
      required: [status, error, request_id]
      properties:
        status:
          type: string
          enum: [Failed]
          example: Failed
        error:
          type: object
          required: [error_code, message]
          properties:
            error_code:
              type: string
              pattern: '^ERROR-[0-9]{3}$'
              example: ERROR-006
            message:
              type: string
              example: "요청한 리소스를 찾을 수 없습니다."
            details:
              type: [object, "null"]
              additionalProperties: true
        request_id:
          type: string
          example: "req_20260714_9999"
    ValidationErrorItem:
      type: object
      description: 필드 단위 검증 오류 항목.
      required: [field, reason]
      properties:
        field:
          type: string
          example: deposit
        reason:
          type: string
          example: "0보다 큰 정수여야 합니다."
    # ======================================================================
    # 2. 공통 Enum (Common Enum) - 14종 (개발설계보고서 10.2절, 부록 C 및 관련 문서 기준)
    # ======================================================================
...
    BlockchainStatus:
      type: string
      description: 온체인 기록 상태(개발설계보고서 10.2절). Confirmed라도 법적 효력을 의미하지 않는다(Frontend_UIUX_명세서 10.4절).
      enum: [NotRequested, Pending, Confirmed, Failed]
      example: Confirmed
...
    UserRole:
      type: string
      description: 사용자 역할(개발설계보고서 13.5절, Frontend_UIUX_명세서 3장). verifier는 advisor 내 서브 권한이다.
      enum: [tenant, landlord, advisor, hug_admin, system_admin, verifier]
      example: tenant
...
    User:
      type: object
      description: 사용자와 역할(개발설계보고서 부록 B user_id).
      required: [user_id, role]
      properties:
        user_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        role:
          $ref: '#/components/schemas/UserRole'
        display_name:
          type: string
          example: "홍길동"
        email:
          type: [string, "null"]
          format: email
        is_active:
          type: boolean
          default: true
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
        last_login_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
...
    BlockchainTransaction:
      type: object
      description: 블록체인 트랜잭션 로그 엔터티(개발설계보고서 부록 B blockchain_tx_id).
      required: [blockchain_tx_id, event_type, reference_id, blockchain_status]
      properties:
        blockchain_tx_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        event_type:
          type: string
          description: 10.3절 공통 이벤트 중 온체인 기록 대상 이벤트.
          example: VerificationCompleted
        reference_id:
          $ref: '#/components/schemas/UUID'
        chain_id:
          type: integer
          example: 80002
          description: Polygon Amoy 테스트넷 chain_id(확인 필요, Blockchain_설계서_260714.md에서 확정).
        contract_address:
          type: [string, "null"]
        tx_hash:
          type: [string, "null"]
          readOnly: true
        result_hash:
          type: string
          readOnly: true
        blockchain_status:
          $ref: '#/components/schemas/BlockchainStatus'
        is_mock:
          type: boolean
          default: false
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
        confirmed_at:
          type: [string, "null"]
          format: date-time
          readOnly: true
...
    AnchorRequest:
      type: object
      description: 온체인 기록 요청 바디(POST /blockchain/anchor).
      required: [event_type, reference_id, result_hash]
      properties:
        event_type:
          type: string
          description: 10.3절 공통 이벤트 중 온체인 기록 대상.
          example: VerificationCompleted
        reference_id:
          $ref: '#/components/schemas/UUID'
        result_hash:
          type: string
          example: "sha256:2c624232cdd221771294dfbb310aca000a0df6ac8b66b696d90ef06fdefb64a"
        model_version:
          type: [string, "null"]
          description: ML 결과를 함께 공증하는 경우에만 사용.
    AnchorResponse:
      type: object
      description: 온체인 기록 요청 접수 응답(BlockchainStatus=Pending으로 시작).
      required: [blockchain_tx_id, blockchain_status]
      properties:
        blockchain_tx_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        chain_id:
          type: integer
          example: 80002
        contract_address:
          type: [string, "null"]
        tx_hash:
          type: [string, "null"]
          readOnly: true
        blockchain_status:
          $ref: '#/components/schemas/BlockchainStatus'
        is_mock:
          type: boolean
          default: false
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: tenant01@example.com
        password:
          type: string
          format: password
          writeOnly: true
          example: "P@ssw0rd!"
    LoginResponse:
      type: object
      required: [status, data, request_id]
      properties:
        status:
          type: string
          enum: [Success]
        data:
          type: object
          required: [access_token, token_type, user]
          properties:
            access_token:
              type: string
              readOnly: true
            token_type:
              type: string
              enum: [Bearer]
            expires_in:
              type: integer
              description: 초 단위 만료 시간.
              example: 3600
            user:
              $ref: '#/components/schemas/User'
        request_id:
          type: string
...